# src/openapi/openapi.yml
openapi: 3.0.3
info:
  title: CiberShield Colombia API
  version: 1.0.0
  description: >
    Plataforma SOA de monitoreo de ciberseguridad para PYMEs colombianas.
    Centraliza autenticación, escaneo de activos, análisis de riesgos,
    alertas, cumplimiento normativo y generación de reportes ejecutivos.
servers:
  - url: https://cibershield-production.up.railway.app
    description: Production
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Alert:
      type: object
      required: [id, company_id, type, asset, risk_level, summary, read, created_at]
      properties:
        id:
          type: integer
        company_id:
          type: integer
        type:
          type: string
        asset:
          type: string
        risk_level:
          type: string
          enum: [critico, alto, medio, bajo]
        summary:
          type: string
        read:
          type: boolean
        created_at:
          type: string
          format: date-time
      example:
        id: 1
        company_id: 1
        type: "vulnerability_detected"
        asset: "181.58.100.42"
        risk_level: "alto"
        summary: "Se detectó vulnerabilidad CVE-2023-44487 (CVSS 7.5) en el host 181.58.100.42"
        read: false
        created_at: "2026-04-28T10:30:00.000Z"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
      example:
        error: "Descripción del error ocurrido"
security:
  - bearerAuth: []
paths:
...
  /api/san/alerts/{id}/read:
    put:
      tags: [Alerts]
      summary: Marcar alerta como leída
      description: Actualiza el estado de lectura de una alerta específica.
      operationId: san.alerts.markRead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Alerta marcada como leída
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
              example:
                id: 1
                company_id: 1
                type: "vulnerability_detected"
                asset: "181.58.100.42"
                risk_level: "alto"
                summary: "Se detectó CVE-2023-44487 (CVSS 7.5) en el host 181.58.100.42"
                read: true
                created_at: "2026-04-28T10:30:00.000Z"
        '400':
          description: ID inválido o alerta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Alert not found"
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid or expired token"
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Error al actualizar la alerta"
  /api/san/alerts/read-all:
    put:
      tags: [Alerts]
      summary: Marcar todas las alertas como leídas
      description: Marca como leídas todas las alertas del usuario autenticado.
      operationId: san.alerts.markAllRead
      responses:
        '200':
          description: Todas las alertas marcadas como leídas
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                    description: Número de alertas actualizadas
              example:
                updated: 5
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid or expired token"
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Error al actualizar alertas"
...
  /api/san/alerts/mock:
    post:
      tags: [Alerts]
      summary: Generar alertas de prueba aleatorias
      description: Genera y persiste alertas de seguridad aleatorias para la empresa del usuario autenticado. Útil para pruebas y demostraciones.
      operationId: san.alerts.generateMock
      parameters:
        - in: query
          name: count
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: Cantidad de alertas a generar (1-20)
          example: 5
      responses:
        '201':
          description: Alertas generadas exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated:
                    type: integer
                    description: Cantidad de alertas creadas
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
              example:
                generated: 3
                alerts:
                  - id: 10
                    company_id: 1
                    type: "vulnerability_detected"
                    asset: "181.58.100.42"
                    risk_level: "critico"
                    summary: "Se detectó vulnerabilidad CVE-2021-44228 (CVSS 10.0) — Log4Shell presente en 181.58.100.42"
                    read: false
                    created_at: "2026-04-28T10:00:00.000Z"
                  - id: 11
                    company_id: 1
                    type: "new_port_open"
                    asset: "190.24.135.77"
                    risk_level: "alto"
                    summary: "Puerto 3389 (RDP) detectado expuesto a internet en 190.24.135.77"
                    read: false
                    created_at: "2026-04-28T10:00:01.000Z"
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid or expired token"
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Error al generar alertas de prueba"

  # ── SCN — Cumplimiento Normativo ───────────────────────────────────────────
...